test : UketEventRandomUtil 리팩토링
