+++ src/test/kotlin/uket/domain/uketEvent/EntryGroupServiceTest.kt
@@ -5,6 +5,7 @@ import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
 import io.mockk.every
 import io.mockk.mockk
+import uket.domain.uketEvent.util.UketEventRandomUtil
 import uket.domain.uketevent.entity.EntryGroup
 import uket.domain.uketevent.entity.UketEvent
 import uket.domain.uketevent.entity.UketEventRound
@@ -43,18 +44,17 @@ class EntryGroupServiceTest :
     companion object {
         private fun setDB(now: LocalDateTime): Triple<UketEvent, UketEventRound, List<EntryGroup>> {
             val uketEventRounds =
-                UketEventRandomUtil.createUketEventsRoundsWithDate(
+                UketEventRandomUtil.createUketEventRounds(
                     listOf(now.plusDays(7)),
                     now.minusDays(2),
                     now.plusDays(2),
                 )
 
             val uketEvent = UketEventRandomUtil.createUketEvent(
-                uketEventRounds,
-                listOf()
+                uketEventRounds
             )
 
-            val entryGroups = UketEventRandomUtil.createEntryGroupWithTime(
+            val entryGroups = UketEventRandomUtil.createEntryGroup(
                 uketEventRounds[0],
                 listOf(now.minusHours(1), now, now.plusHours(1))
             )
