+++ src/test/kotlin/uket/domain/uketEvent/UketEventRoundServiceTest.kt
@@ -5,6 +5,7 @@ import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
 import io.mockk.every
 import io.mockk.mockk
+import uket.domain.uketEvent.util.UketEventRandomUtil
 import uket.domain.uketevent.entity.UketEvent
 import uket.domain.uketevent.entity.UketEventRound
 import uket.domain.uketevent.repository.UketEventRoundRepository
@@ -58,7 +59,28 @@ class UketEventRoundServiceTest :
             val now = LocalDateTime.now()
 
             val uketEventRounds =
-                UketEventRandomUtil.createUketEventsRoundsWithDate(
+                UketEventRandomUtil.createUketEventRounds(
+                    listOf(
+                        now.plusDays(10),
+                        now.plusDays(8)
+                    ),
+                    now.minusDays(3),
+                    now.plusDays(2)
+                )
+
+            val uketEvent =
+                UketEventRandomUtil.createUketEvent(
+                    uketEventRounds
+                )
+
+            return Pair(uketEvent.id, listOf(uketEventRounds[0], uketEventRounds[1]))
+        }
+
+        fun setDB2(): Pair<List<UketEvent>, List<UketEventRound>> {
+            val now = LocalDateTime.now()
+
+            val uketEventRounds =
+                UketEventRandomUtil.createUketEventRounds(
                     listOf(
                         now.plusDays(10),
                         now.plusDays(8)
@@ -70,36 +92,13 @@ class UketEventRoundServiceTest :
             val uketEvent =
                 UketEventRandomUtil.createUketEvent(
                     uketEventRounds,
-                    listOf()
-                )
-
-            return Pair(uketEvent.id, listOf(uketEventRounds[0], uketEventRounds[1]))
-        }
-
-        fun setDB2(): Pair<List<UketEvent>, List<UketEventRound>> {
-            val now = LocalDateTime.now()
-
-            val uketEventRounds =
-                UketEventRandomUtil.createUketEventsRoundsWithDate(
-                    listOf(
-                        now.plusDays(10),
-                        now.plusDays(8)
-                    ),
-                    now.minusDays(3),
-                    now.plusDays(2)
-                )
-
-            val uketEvent =
-                UketEventRandomUtil.createUketEventWithNameAndId(
-                    now.minusDays(3),
-                    now.plusDays(2),
+                    listOf(),
                     "eventA",
                     1L,
-                    uketEventRounds,
                 )
 
             val uketEventRounds2 =
-                UketEventRandomUtil.createUketEventsRoundsWithDate(
+                UketEventRandomUtil.createUketEventRounds(
                     listOf(
                         now.plusDays(10),
                         now.plusDays(8)
@@ -109,12 +108,11 @@ class UketEventRoundServiceTest :
                 )
 
             val uketEvent2 =
-                UketEventRandomUtil.createUketEventWithNameAndId(
-                    now.minusDays(3),
-                    now.plusDays(2),
-                    "eventB",
-                    2L,
+                UketEventRandomUtil.createUketEvent(
                     uketEventRounds,
+                    listOf(),
+                    "eventB",
+                    2L
                 )
 
             val rounds = mutableListOf<UketEventRound>()
