+++ src/test/kotlin/uket/domain/uketEvent/UketEventServiceTest.kt
@@ -5,6 +5,7 @@ import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
 import io.mockk.every
 import io.mockk.mockk
+import uket.domain.uketEvent.util.UketEventRandomUtil
 import uket.domain.uketevent.entity.UketEvent
 import uket.domain.uketevent.repository.UketEventRepository
 import uket.domain.uketevent.service.UketEventService
@@ -39,23 +40,23 @@ class UketEventServiceTest :
     }) {
     companion object {
         fun setDB(now: LocalDateTime): List<UketEvent> {
-            val uketEventRounds1 = UketEventRandomUtil.createUketEventsRoundsWithDate(
+            val uketEventRounds1 = UketEventRandomUtil.createUketEventRounds(
                 listOf(now, now.plusDays(1), now.plusDays(2), now.plusDays(3), now.plusDays(4)),
                 now.minusDays(4), now.minusDays(3)
             )
-            val uketEvent1 = UketEventRandomUtil.createUketEvent(uketEventRounds1, listOf())
+            val uketEvent1 = UketEventRandomUtil.createUketEvent(uketEventRounds1)
 
-            val uketEventRounds2 = UketEventRandomUtil.createUketEventsRoundsWithDate(
+            val uketEventRounds2 = UketEventRandomUtil.createUketEventRounds(
                 listOf(now.plusDays(1), now.plusDays(2), now.plusDays(3), now.plusDays(4)),
                 now.minusDays(4), now.minusDays(3)
             )
-            val uketEvent2 = UketEventRandomUtil.createUketEvent(uketEventRounds2, listOf())
+            val uketEvent2 = UketEventRandomUtil.createUketEvent(uketEventRounds2)
 
-            val uketEventRounds3 = UketEventRandomUtil.createUketEventsRoundsWithDate(
+            val uketEventRounds3 = UketEventRandomUtil.createUketEventRounds(
                 listOf(now.minusDays(1)),
                 now.minusDays(4), now.minusDays(3)
             )
-            val uketEvent3 = UketEventRandomUtil.createUketEvent(uketEventRounds3, listOf())
+            val uketEvent3 = UketEventRandomUtil.createUketEvent(uketEventRounds3)
 
             return listOf(uketEvent1, uketEvent2, uketEvent3)
         }
